backend/requirements.txt
# FastAPI and ASGI server
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Google Cloud services (optional - for full functionality)
google-cloud-aiplatform>=1.38.0
google-cloud-documentai>=2.21.0
google-cloud-vision>=3.4.0
google-cloud-translate>=3.12.0
google-cloud-texttospeech>=2.16.0
google-cloud-speech>=2.21.0
google-cloud-storage>=2.10.0
google-cloud-firestore>=2.13.0
google-cloud-dlp>=3.12.0
google-cloud-workflows>=1.12.0
google-cloud-tasks>=2.14.0
google-cloud-pubsub>=2.18.0
google-cloud-monitoring>=2.16.0
google-cloud-logging>=3.8.0
google-cloud-error-reporting>=1.12.0
google-cloud-bigquery>=3.13.0

# CrewAI and LangChain (optional - for AI functionality)
# crewai>=0.22.0
# langchain>=0.1.0
# langchain-google-vertexai>=0.0.6

# Firebase Admin SDK (optional)
firebase-admin>=6.2.0

# Data processing and validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# HTTP client and utilities
httpx>=0.25.2
aiofiles>=23.2.1
python-dotenv>=1.0.0
aiohttp>=3.9.0

# AI Agent Services
google-generativeai>=0.3.0

# Document processing (optional)
PyPDF2>=3.0.1
python-docx>=1.1.0
Pillow>=10.1.0
opencv-python>=4.8.0
numpy>=1.24.0
python-magic>=0.4.27
reportlab>=4.0.0

# Database and caching (optional)
# redis>=5.0.1
# sqlalchemy>=2.0.23

# Note: Development dependencies are in requirements-dev.txt

# Monitoring and logging
structlog>=23.2.0
sentry-sdk[fastapi]>=1.38.0